backend/requirements.txt
backports.functools-lru-cache==1.6.4
bcrypt==4.0.1
Brotli==1.0.9
certifi==2022.9.24
cffi==1.15.1
charset-normalizer==2.1.1
click==8.0.1
colorama==0.4.4
cryptography==38.0.3
dash-core-components==2.0.0
dash-html-components==2.0.0
dash-table==5.0.0
debugpy==1.6.3
decorator==5.1.1
dnspython==2.2.1
ecdsa==0.18.0
email-validator==1.3.0
entrypoints==0.4
fastapi==0.87.0
Flask==2.0.1
Flask-Compress==1.10.1
greenlet==2.0.1
h11==0.12.0
idna==3.4
importlib-metadata==4.8.1
importlib-resources==5.10.0
ipykernel==6.15.0
ipython==7.33.0
itsdangerous==2.0.1
jedi==0.18.2
Jinja2==3.0.1
jupyter_client==7.4.7
jupyter_core==4.11.2
Mako==1.2.4
MarkupSafe==2.0.1
matplotlib-inline==0.1.6
nest-asyncio==1.5.6
packaging==21.3
parso==0.8.3
passlib==1.7.4
pickleshare==0.7.5
pip==22.2.2
plotly==5.3.1
prompt-toolkit==3.0.33
psutil==5.9.0
psycopg2-binary
pyasn1==0.4.8
pycparser==2.21
pydantic==1.10.2
Pygments==2.13.0
pyparsing==3.0.9
python-dateutil==2.8.2
python-jose==3.3.0
python-multipart==0.0.5
pyzmq==23.2.0
requests==2.28.1
rsa==4.9
setuptools==65.5.0
six==1.16.0
sniffio==1.3.0
SQLAlchemy==1.4.44
starlette==0.21.0
tenacity==8.0.1
tornado==6.2
traitlets==5.5.0
typing_extensions==4.4.0
urllib3==1.26.13
uvicorn==0.19.0
wcwidth==0.2.5
Werkzeug==2.0.1
wheel==0.37.1
wincertstore==0.2
zipp==3.6.0
fastapi-jwt-auth==0.5.0
google-auth==2.16.0
redis==4.4.2